api/pyproject.toml
[project]
name = "chat-room-api"
version = "0.1.0"
description = "Chat Room Backend API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "python-multipart>=0.0.6",
    "Pillow>=10.0.0",
    "bcrypt>=4.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
